feat(runtime): enhance parallel execution support and logging
- Implement new debug loggers for detailed execution flow tracking.
- Refactor workflow loader and compiler for better maintainability.
- Add test cases to verify the functionality and prevent regressions.
- Introduce performance tweaks for better parallel execution support.
- Adjust thread pool naming conventions for clarity.
- Fix a bug where node definitions were not being deleted correctly.
- BREAKING CHANGE: The execution flow and logging mechanisms have been reworked,
which may affect existing clients relying on the previous behavior.

+++ innospots-base/src/main/java/io/innospots/base/script/ScriptExecutorManager.java
@@ -69,6 +69,9 @@ public class ScriptExecutorManager {
 
     protected String packageName;
 
+
+    private static boolean retainSource = true;
+
     protected Map<String, IScriptExecutor> executors = new HashMap<>();
 
     /**
@@ -149,7 +152,9 @@ public class ScriptExecutorManager {
                 compiler.compile();
                 //reload();
                 sourceBuilder.clear();
-                sourceBuilder.deleteSourceFile();
+                if(!retainSource){
+                    sourceBuilder.deleteSourceFile();
+                }
                 logger.info("engine class file write complete, classFile:{} , loaded executor size:{}", className(), executors.size());
                 return true;
             } catch (IOException e) {
@@ -188,6 +193,10 @@ public class ScriptExecutorManager {
         System.setProperty(SOURCE_PATH_EVN, sourcePath);
     }
 
+    public static void setRetainSource(boolean retainSource){
+        ScriptExecutorManager.retainSource = retainSource;
+    }
+
     public static String getClassPath() {
         return System.getProperty(CLASS_PATH_ENV);
     }
@@ -203,8 +212,8 @@ public class ScriptExecutorManager {
      */
     public void clear() {
         File clsFile = new File(classPath().toFile(), className().replace(".", File.separator) + ".class");
-        logger.debug("remove class file:{}", clsFile.getPath());
         if (clsFile.exists()) {
+            logger.debug("remove class file:{}", clsFile.getPath());
             clsFile.delete();
         }
         sourceBuilder.deleteSourceFile();
